Remove ctags and its manpage from the emacs build
In addition to its etags executable, emacs now installs a ctags executable.
This stomps over exuberant-ctags, and makes it difficult to have emacs and
vim installed and use tags for both.

This follows what MacPorts does and removes bin/ctags and its manpage from
the build before linking.

+++ Library/Formula/emacs.rb
@@ -6,13 +6,17 @@ class Emacs < Formula
   mirror 'http://ftp.gnu.org/pub/gnu/emacs/emacs-24.2.tar.bz2'
   sha1 '38e8fbc9573b70a123358b155cf55c274b5a56cf'
 
-  conflicts_with 'ctags',
-    :because => 'Both emacs and ctags install a `ctags` executable.'
-
   option "cocoa", "Build a Cocoa version of emacs"
   option "srgb", "Enable sRGB colors in the Cocoa version of emacs"
   option "with-x", "Include X11 support"
   option "use-git-head", "Use Savannah git mirror for HEAD builds"
+  option "keep-ctags", "Don't remove the ctags executable that emacs provides"
+
+  if build.include? "keep-ctags"
+    conflicts_with 'ctags',
+      :because => 'Both emacs and ctags install a `ctags` executable.'
+  end
+
 
   if build.include? "use-git-head"
     head 'http://git.sv.gnu.org/r/emacs.git'
@@ -65,6 +69,13 @@ class Emacs < Formula
       system "make install"
       prefix.install "nextstep/Emacs.app"
 
+      # Follow MacPorts and don't install ctags from emacs. This allows vim
+      # and emacs and ctags to play together without violence.
+      unless build.include? "keep-ctags"
+        (bin/"ctags").unlink
+        (share/man/man1/"ctags.1.gz").unlink
+      end
+
       # Replace the symlink with one that avoids starting Cocoa.
       (bin/"emacs").unlink # Kill the existing symlink
       (bin/"emacs").write <<-EOS.undent
@@ -87,6 +98,13 @@ class Emacs < Formula
       system "./configure", *args
       system "make"
       system "make install"
+
+      # Follow MacPorts and don't install ctags from emacs. This allows vim
+      # and emacs and ctags to play together without violence.
+      unless build.include? "keep-ctags"
+        (bin/"ctags").unlink
+        (share/man/man1/"ctags.1.gz").unlink
+      end
     end
   end
